% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_col}
\alias{add_col}
\title{Add a new column to a data frame
Add a new column defined by the user and put in on the existing data frame}
\usage{
add_col(df, name = "y", value, where = ncol(df) + 1)
}
\arguments{
\item{df}{A data frame}

\item{name}{Name of the new column}

\item{value}{Value of the new column}

\item{where}{Where the new column will be located}
}
\value{
A data frame
}
\description{
Add a new column to a data frame
Add a new column defined by the user and put in on the existing data frame
}
\examples{
df <- data.frame(x = 1)
add_col(df, "x", 2, where = 2)
}
