% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quotes.R
\name{add_quotes}
\alias{add_quotes}
\title{Put double quote around text}
\usage{
add_quotes(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
Enclose text in quotes as may be useful for sending commands to Linux terminal,
e.g. if a path contains spaces or ampersands
}
\examples{
add_quotes("~/badly & named directory")
}
