% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitracks.R
\name{addlist.multitracks}
\alias{addlist.multitracks}
\title{Addlist method for adding new track to \code{multitracks} object}
\usage{
\method{addlist}{multitracks}(existing_list, new_line)
}
\arguments{
\item{existing_list}{This is object of a class \code{multitracks} which holds
multiple lines or only one line so far}

\item{new_line}{object of a class \code{multitracks} which we wish to add}
}
\value{
Does not return new object, instead it adds new_line to existing_list
}
\description{
Addlist method for adding new track to \code{multitracks} object
}
\examples{
\dontrun{
addlist(my_tracks,line2)
}
}
