% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitracks.R
\name{addlist}
\alias{addlist}
\title{Generic function for adding new track of class \code{mytrack} to object of class
\code{multitracks}}
\usage{
addlist(existing_list, new_line)
}
\arguments{
\item{existing_list}{This is object of a class \code{multitracks} which holds
multiple lines or only one line so far}

\item{new_line}{object of a class \code{mytrack} which we wish to add}
}
\value{
Does not return new object, instead it adds new_line to existing_list
}
\description{
Generic function for adding new track of class \code{mytrack} to object of class
\code{multitracks}
}
