% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPercent.R
\name{addPercent}
\alias{addPercent}
\title{Add percent to factor}
\usage{
addPercent(x)
}
\arguments{
\item{x}{factor (numeric)}
}
\value{
factor
}
\description{
Add percent to factor
}
\examples{
addPercent(iris$Sepal.Length)
}

