% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_factors.R
\name{any_factors}
\alias{any_factors}
\title{Check a dataframe for factors!}
\usage{
any_factors(a)
}
\arguments{
\item{a}{dataframe}
}
\value{
vector
}
\description{
Factors are sneaky little buggers. The stealth-factor is a constant source of annoyance. Use this to check if your
dataframe contains any factors you don't expect.
}
\examples{
#' any_factors(iris)
#' any_factors(mtcars)
}

