% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_is_levels.R
\name{as_is_levels}
\alias{as_is_levels}
\title{as is levels}
\usage{
as_is_levels(x)
}
\arguments{
\item{character}{to be converte, or factor to be modified}
}
\value{
factor with unrodered levels
}
\description{
returns a factor with levels in their original order
Example:
pollo <- c("rooster", "chicken", "hen")
as_is_levels(pollo)
[1] rooster chicken hen
Levels: rooster chicken hen
}
