% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_two.R
\name{average_two}
\alias{average_two}
\title{Averages two numbers}
\usage{
average_two(a, b)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
number
}
\description{
Averages two numbers.
}
\examples{

}
