% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{change de the order of the factors backwards}
\usage{
backward(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
change de the order of the factors backwards
}
\examples{
backward(iris$Species)
}
