% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_or_factor.R
\name{char_not_fact}
\alias{char_not_fact}
\title{character or factor}
\usage{
char_not_fact(x)
}
\arguments{
\item{vector}{to be assess}
}
\value{
whether it should be character or factor
}
\description{
rcompares the lenght of levels with the leght of the factor, if equal, it suggest it should be character, if not, factor.
}
\details{
Example:

pollos<- c("gallo", "gallo", "pollito", "gallina", "pollito")
char_not_fact(pollos)
[1] "it should be a factor"

pollo <- c("rooster", "chicken", "hen")
char_not_fact(pollo)
[1] "that should be character"
}
