% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{character}
\alias{character}
\title{Detect factors that should be character because # unique values = length}
\usage{
character(n)
}
\arguments{
\item{n}{factor}
}
\value{
tbl_df
}
\description{
Detect factors that should be character because # unique values = length
}

