% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_factor.R
\name{check_factor}
\alias{check_factor}
\title{Check whether the input is a factor}
\usage{
check_factor(a)
}
\arguments{
\item{a}{input}
}
\value{
true: if it is a factor; false: if it is not a factor
}
\description{
This is a function which is used to double check whether the input is a factor
or not and will return the true or false judgement
}
\examples{
check_factor(factor(c("hits","hits","your", "eyeballs")))
check_factor(factor(c("but","integer", "where it", "counts", "but")))
}
