% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_factor_character.R
\name{check_factor_character}
\alias{check_factor_character}
\title{Detects factors that should be characters}
\usage{
check_factor_character(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
Detects factors that should be character, because
# unique values = length
}
\examples{
#' check_factor_character(iris$Species)
}
