% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_factors.R
\name{check_factors}
\alias{check_factors}
\title{Check a dataframe for factors}
\usage{
check_factors(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
vector
}
\description{
We can use this function to check if your dataframe contains any factors you
don't want to use.
}
\examples{
check_factors(iris)
}
