% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_factor.R
\name{chr_factor}
\alias{chr_factor}
\title{Factor should be character}
\usage{
chr_factor(a)
}
\arguments{
\item{a}{factor}
}
\value{
logical
}
\description{
Check to see if a factor should be character. The function checks whether the length of the
factor is equal to the length of the unique values.
}
\examples{
#' chr_factor(factor(c('BumbleBees', 'Bees', 'Bears'))
}

