% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuttable.R
\name{cuttable}
\alias{cuttable}
\title{Make a distribution table for a factor}
\usage{
cuttable(x, a)
}
\arguments{
\item{x}{factor (numeric)}

\item{a}{numeric vector}
}
\value{
tbl_df
}
\description{
Make a distribution table for a factor
}
\examples{
cuttable(iris$Sepal.Length,3)
}

