% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_order.R
\name{desc_order}
\alias{desc_order}
\title{Reorders a vector in descending format based on the levels}
\usage{
desc_order(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Reorders a vector in descending format based on the levels
}

