% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_reord.R
\name{desc_reord}
\alias{desc_reord}
\title{Reorder levels of argument in descending order}
\usage{
desc_reord(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
My version of reorder that uses dplyr::desc() to reorder
factor levels in descending order
}
\examples{
#' desc_reord(iris$Species)
}

