% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fac.R
\name{detect_fac}
\alias{detect_fac}
\title{Detect factors that should be characters}
\usage{
detect_fac(a)
}
\arguments{
\item{a}{list of factors}
}
\value{
a list of characters
}
\description{
This function will detect a list of factors that should be characters because its number of unique levels
is actually the same as the length of the list. If the list does have some levels that appear more than
once, then this function will not do anything.
}
\examples{
detect_fac(factor(c("a","b","c")))
}

