% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fct.R
\name{detect_fct}
\alias{detect_fct}
\title{Detect whether a factor is truly factor or character}
\usage{
detect_fct(x)
}
\arguments{
\item{x}{factor}
}
\value{
boolean; \code{TRUE} if input should be considered as factor,
   \code{FALSE} if input should be considered as character
}
\description{
If the number of unique values equal to the length of input, we consider the input as factor; otherwise,
we consider it as character
}
\examples{
a <- factor(c("high", "high", "low"))
b <- factor(c("high", "low", "medium"))

detect_fct(a)

detect_fct(b)

}
