% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read.R
\name{df_read}
\alias{df_read}
\title{Read data frame}
\usage{
df_read(dfname, txtfilename = "plain.txt")
}
\arguments{
\item{dfname}{path for the data frame}
}
\description{
Read data frames to plain text delimited files while retaining factor levels
}
