% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_write.R
\name{df_write}
\alias{df_write}
\title{Write data frame to plain text delimited files}
\usage{
df_write(df, df_name, file_path = "plain.txt")
}
\arguments{
\item{df}{a data frame to be stored.}

\item{df_file_name}{txt file name to store the file of the data frame.}
}
\description{
Write data frames to plain text delimited files while retaining factor levels
}
