% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read.R, R/df_write.R
\name{df_read}
\alias{df_read}
\alias{df_write}
\title{Write and read data frame to/from text files}
\usage{
df_read(filename)

df_write(df, filename)
}
\arguments{
\item{filename}{A string, which is the name of the data frame .txt file, the input text file need to save the same folder with the R script folder, if it's not a ".txt" file, this function will come to an error.}

\item{df}{data frame}
}
\value{
df_read() returns a data frame read from a txt file

df_write() returns nothing
}
\description{
This function is used to write and read data frames to plain text delimited files while retaining factor levels.
}
