% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readandwrite.R
\name{dfread}
\alias{dfread}
\title{read data frames from plain text delimited}
\usage{
readdataframe(file, level)
}
\arguments{
\item{level}{is where to store the file}

\item{dataframe}{is dataframe}
}
\description{
read data frame from plain text delimited files
}
