% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/die_factors.R
\name{die_factors}
\alias{die_factors}
\title{Kill factors in dataframe}
\usage{
die_factors(x, targets = NULL)
}
\arguments{
\item{x}{}

\item{targets}{}
}
\value{
tbl_df
}
\description{
Converts all factors in a given dataframe to characters.
}
\examples{
die_factors(iris), die_factors(iris, "Species"), die_factors(iris, 5)
}

