% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks.R
\name{distance.mytrack}
\alias{distance.mytrack}
\title{Distance method for \code{mytrack} class objects}
\usage{
\method{distance}{mytrack}(obj, status)
}
\arguments{
\item{obj}{Object of class \code{mytrack}}

\item{status}{TRUE for using S2 for distance calculation, or FALSE for not using it}
}
\value{
Distance in meters
}
\description{
Distance method for \code{mytrack} class objects
}
\examples{
\dontrun{
distance.mytrack(obj1,TRUE)
}
}
