% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetrack.R
\name{distance}
\alias{distance}
\alias{distance,track-method}
\title{Distance generic}
\usage{
distance(obj, status)

\S4method{distance}{track}(obj, status)
}
\arguments{
\item{obj}{Object of class track}

\item{status}{Object of class logical}
}
\description{
Distance generic
}
\section{Methods (by class)}{
\itemize{
\item \code{track}: Method for calculating distance of a track
}}

