% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{distance2.mytrack}
\alias{distance2.mytrack}
\title{Distance method for "mytrack" class objects}
\usage{
\method{distance2}{mytrack}(obj, status)
}
\arguments{
\item{obj}{Object of class "mytrack"}

\item{status}{TRUE for using S2 for distance calculation, or FALSE for not using it}
}
\value{
Distance in meters
}
\description{
Distance method for "mytrack" class objects
}
\examples{
\dontrun{
distance2.mytrack(obj1,TRUE)
}
}
