% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{distance2}
\alias{distance2}
\title{Generic distance function}
\usage{
distance2(obj, status)
}
\arguments{
\item{obj}{Object of class "mytrack"}

\item{status}{TRUE for using S2 for distance calculation, or FALSE for not using it}
}
\value{
Distance in meters
}
\description{
Generic distance function
}
