% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreorder.R
\name{dreorder}
\alias{dreorder}
\title{Sets levels of factors to the order in which they appear in the data}
\usage{
dreorder(a, b)
}
\arguments{
\item{a}{a list of factors}

\item{b}{a vector of the same length as a, whose subset of values for each unique level
of a determines the eventual order of that level.}
}
\value{
same list of factors but with descending order with respect to the measure b
}
\description{
Sets levels of factors to the order in which they appear in the data
}
\examples{
x <- data.frame(homework = factor(c('560', '547', '545')), length = c(2, 3, 1))
with(dreorder(homework, length), data = x)
}

