% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_Standings_data.R
\docType{data}
\name{Driver_Standings_data}
\alias{Driver_Standings_data}
\title{Driver_Standings_data}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{Driver}{Driver's Name}
\item{PTS}{Points}
...
}
}
\source{
\url{https://www.formula1.com}
}
\usage{
Driver_Standings_data
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
