% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_keep.R
\name{f_keep}
\alias{f_keep}
\title{Print factor levels as they are}
\usage{
f_keep(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor with factor levels in order in which levels appear in oject.
}
\description{
Print factor levels as they are
}
\examples{
 x <- factor(c("He", "Came", "To", "Sch"))
 f_keep(x)

}

