% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_nchar.R
\name{f_nchar}
\alias{f_nchar}
\title{Number of Characters in the level name}
\usage{
f_nchar(x)
}
\arguments{
\item{x}{factor}
}
\value{
data frame with levels and number of characters
}
\description{
Number of Characters in the level name
}
\examples{
Subjects <- factor(c("Mathematics", "Statistics", "English", "French"))
         f_nchar(Subjects)


}

