% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac.R
\name{fac}
\alias{fac}
\title{Factors}
\usage{
fac(x)
}
\arguments{
\item{x}{}
}
\value{
factor
}
\description{
Turn a vector to a factor and set levels to the order in which they appear in the data
}
\examples{
fac(c(3,2,5,1,1,1,6))
}

