% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_detect.r
\name{fact_detect}
\alias{fact_detect}
\title{Detech factor should be character or not}
\usage{
fact_detect(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
Tells whether the inputted factor should be a character
by comparing levels and factor (True means should be character)
and false means should not be character
}
\examples{
fact_detect(iris$Species)
}
