% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_orig.r
\name{fact_orig}
\alias{fact_orig}
\title{Gives the levels in the same order as input}
\usage{
fact_orig(x)
}
\arguments{
\item{x}{factor or Character}
}
\value{
Factor
}
\description{
Returns the factor with levels same as that of the
input character vector
}
\examples{
fact_orig(c("This","is","test","examp"))
}
