% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_as_orginal.R
\name{factor_as_orginal}
\alias{factor_as_orginal}
\title{Returns factor levels same as factor levels in Input}
\usage{
factor_as_orginal(x)
}
\arguments{
\item{x}{factor or a character}
}
\value{
factor; return unique levels of input in the order which they appear
}
\description{
Returns factor levels same as factor levels in Input
}
\examples{
data_1 <- c("Asia","Africa","Americas","Europe","Oceania")
factor_as_orginal(data_1)
}
