% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_asis.R
\name{factor_asis}
\alias{factor_asis}
\title{Sets levels to the order in which they appear in the data}
\usage{
factor_asis(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor; return unique levels of input in the order which they appear
}
\description{
Sometimes we want to set levels of a factor to the order which they appear in the data,
this function is able to set levels without duplicates.
}
\examples{

a <- factor(c("high","low","low","high","medium"))
factor_asis(a)
}
