% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_levels.R
\name{factor_levels}
\alias{factor_levels}
\title{Maintain factor levels as is}
\usage{
factor_levels(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
This function turns a character variable into a factor
and keeps the factor levels in the same order that they appear in the dataset.
}
\examples{
(A <- fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)]))
factor_levels(A)
levels(A)
}

