% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Add.R
\name{fAdd}
\alias{fAdd}
\title{Add two numbers}
\usage{
fAdd(p, q)
}
\arguments{
\item{p}{factor}

\item{q}{factor}
}
\value{
factor
}
\description{
Return a number that is the sum of q and q
are the union of the levels of the input factors.
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
