% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcheck.R
\name{fcheck}
\alias{fcheck}
\title{Check if factor type is appropriate}
\usage{
fcheck(x)
}
\arguments{
\item{x}{the factor}
}
\value{
logical
}
\description{
This function checks to see if char type may be more appropriate.
It returns TRUE if it looks okay, and FALSE if the number of levels
is equal to the length of the vector.
}
\examples{
x <- as.factor(c("cat", "dog", "fish", "blob"))
fcheck(x)
fcheck(iris$Species)
}

