% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fclean.R
\name{fclean}
\alias{fclean}
\title{Clean factor levels}
\usage{
fclean(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
For all levels, remove white spaces and convert to title case.
}
\examples{
# Create a factor with messy levels
production <- factor(c("  HIGH production", "MED production  ",  "MED production  ", " low Production"))

# Run it through fclean()
fclean(production)
}
