% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcommon.R
\name{fcommon}
\alias{fcommon}
\title{Do elements of one factor occur in another?}
\usage{
fcommon(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
c vector
}
\description{
Do elements of one factor occur in another?
}
\examples{
x <- c("a","b","c")
y <- c("a","a","c")
fcommon(x,y)
}
