% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fconvert.R
\name{fconvert}
\alias{fconvert}
\title{Convert input to a factor}
\usage{
fconvert(x)
}
\arguments{
\item{x}{vector(s)}
}
\value{
factor(s)
}
\description{
This function is used to encode a vector/vectors as a factor/factors
}
\examples{
fconvert("apple")
fconvert(c("low","medium","high"))
}
