% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fctr_or_chr.R
\name{fctr_or_chr}
\alias{fctr_or_chr}
\title{Factor or character}
\usage{
fctr_or_chr(x)
}
\arguments{
\item{x}{}
}
\value{
logical
}
\description{
Factor or character
}
\examples{
fctr_or_chr(factor(c('a','b','a')))
}

