% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdetect.R
\name{fdetect}
\alias{fdetect}
\title{Detect whether a factor that should be character}
\usage{
fdetect(x)
}
\arguments{
\item{x}{factor}
}
\value{
This funciton returns a boolean value, if the input is a factor then it returns TRUE, if the input is a character then it returns FALSE.
}
\description{
This function is used to detect factors that should be character based on the fact that the length of character equals the number of its unique values, however, factor doesn't have that characteristic.
}
\examples{
FALSE
fdetect(factor(c("a","b","c")))
TRUE
fdetect(factor(c("a","b","c","b")))
}
