% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdiff.R
\name{fdiff}
\alias{fdiff}
\title{Difference between Two Integers}
\usage{
fdiff(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{integer}
}
\value{
absolute difference (integer)
}
\description{
Create a new integer from the difference of two random integers, where new integer
is the absolute value of the difference of two input integers
}
\examples{
 fdiff(iris$Sepal.Length[1],iris$Sepal.Length[2])
}
