% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffactor.R
\name{ffactor}
\alias{ffactor}
\title{Factors}
\usage{
ffactor(a)
}
\arguments{
\item{a}{vector of data, usually taking a small number of distinct values.}
}
\value{
ffactor returns an object of class "factor" which has its levels set to the order as they appear in the data as unique entries
}
\description{
The function ffactor is used to encode a vector as a factor that sets levels to the order in which they appear in the data
}
\examples{
ffactor(c("C", "D", "B", "A"))
}

