% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgap_rw.R
\name{fgap_rw}
\alias{fgap_rw}
\title{function to write and read data frames to plain text delimited files while
retaining factor evels by using writing/reading a companion file'}
\usage{
fgap_rw(dat, na.rm = TRUE)
}
\arguments{
\item{dat}{gapminder data frame}

\item{na.rm}{removes NA (if any)}
}
\value{
txt and rds files and a mini data frame
}
\description{
function to write and read data frames to plain text delimited files while
retaining factor evels by using writing/reading a companion file'
}
\examples{
fgap_rw(gapminder::gapminder)
}

