% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppercase.R
\name{first_upper}
\alias{first_upper}
\title{First letter uppercase}
\usage{
first_upper(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Returns the same factor but with the first letter changred to uppercase
}
\examples{
first_upper(iris$Species)
}
