% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flevel.R
\name{flevel}
\alias{flevel}
\title{Levels and number of levels}
\usage{
flevel(x)
}
\arguments{
\item{x}{factor}
}
\value{
levels

number of levels
}
\description{
This function returns the levels and the number of levels which its factor has.
}
\examples{
flevel(iris$Species)
}
