% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmm.R
\name{fmmm}
\alias{fmmm}
\title{Title: Max Minus Min function}
\usage{
fmmm(x)
}
\arguments{
\item{x}{is an input vector}
}
\value{
number
}
\description{
Title: Max Minus Min function
}
\examples{
fmmm(c(1,2,3,4,5))
}

