% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmultiply.R
\name{fmultiply}
\alias{fmultiply}
\title{Bind two factors}
\usage{
fmultiply(x, y)
}
\arguments{
\item{x}{integer}

\item{y}{integer}
}
\value{
factor
}
\description{
Create a new integer from the product of two random integers, where the new integer
is the multiplication of the two input integers
}
\examples{
fmultiply(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
