% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo_interrobang.R
\name{foo_interrobang}
\alias{foo_interrobang}
\title{Insert an interrobang at the end of factor}
\usage{
foo_interrobang(d)
}
\arguments{
\item{d}{factor}
}
\value{
factor
}
\description{
Insert an interrobang at the end of factor
}
\examples{
foo_interrobang(iris$Species)
}

