% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.R
\docType{package}
\name{foofactors-package}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: combine two factors into one factor}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
