% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.r
\docType{package}
\name{foofactors}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors. Make factors grate again.}
\description{
The first rule about factors is that we don't talk about factors.
The second rule about factors is that we don't talk about factors.
The third rule about factors is a reminder to everyone that the first and second rules about factors are that we don't talk about factors.
And Stuff!
And More Stuff! YEs!
}
