% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forders.R
\name{forders}
\alias{forders}
\title{Preserve order of data factor in levels as it appears or "as is".}
\usage{
forders(d)
}
\arguments{
\item{d}{factor}
}
\value{
factor
}
\description{
Preserve order of data factor in levels as it appears or "as is".
}
