% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frep.R
\name{frep}
\alias{frep}
\title{Repeats a factor n times}
\usage{
frep(x, n = 1)
}
\arguments{
\item{x}{factor}

\item{n}{integer}
}
\value{
factor
}
\description{
Repeats a factor n times
}
\examples{
frep(iris$Species[c(1, 51, 101)], 3)
}

