% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_out.R
\name{freq_out}
\alias{freq_out}
\title{Create a frequency table for facot}
\usage{
freq_out(x)
}
\arguments{
\item{x}{factor}
}
\value{
tbl_df
}
\description{
Create a frequency table for facot
}
\examples{
freq_out(iris$Species)
}

