% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fset.R
\name{fset}
\alias{fset}
\title{Set the levels}
\usage{
fset(x)
}
\arguments{
\item{x}{a factor. It will come to an error if it is not a factor.}
}
\value{
a factor
}
\description{
This function is used to set levels to the order in which they appear in the data
}
\examples{
fset(factor(c("c", "d", "b", "a"))) # Levels: c d b a

}
