% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsetis.R
\name{fsetis}
\alias{fsetis}
\title{Set levels to the order in which they appear in the data,}
\usage{
fsetis(a)
}
\arguments{
\item{a}{factor}
}
\value{
factor
}
\description{
Set levels to the order in which they appear in the data,
}
