% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givenorder.R
\name{givenorder}
\alias{givenorder}
\title{A version of factor() that sets levels to the order in which they appear in the data,
i.e. set the levels “as is”}
\usage{
givenorder(m)
}
\arguments{
\item{m}{factor}
}
\value{
factor
}
\description{
A version of factor() that sets levels to the order in which they appear in the data,
i.e. set the levels “as is”
}

