% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_desc.R
\name{in_desc}
\alias{in_desc}
\title{Reorder in descending}
\usage{
in_desc(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
A version of reorder() using desc() from dplyr.
}
\examples{
in_desc(factor(c('a', 'b', 'c')))

}

