% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Insert columns into the dataframes}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{first dataframe}

\item{y}{second one which will be inserted}

\item{where}{the number of column before which we will insert}
}
\value{
this is a dataframe
}
\description{
Insert columns into the dataframes
}
\examples{

x <- data.frame(a = 3, b = 4, c = 5)
y <- data.frame(X = 1, Y = 2)
where <- 2
insert_into(x, y)


}
