% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_char.R
\name{is_char}
\alias{is_char}
\title{Factor is charactor}
\usage{
is_char(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
Detect factors that should be character because the
number of unique values equals its length.
}
\examples{
is_char(factor(c('Bird', 'Plane', 'Superman')))
}

