% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_char_factor.R
\name{is_char_factor}
\alias{is_char_factor}
\title{Detects factors that should be characters}
\usage{
is_char_factor(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
Detects factors that should be character, because
# unique values = length
}
\examples{
#' is_char_factor(iris$Species)
}

