% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.R
\name{is_even}
\alias{is_even}
\title{Check if number(s) is/are even}
\usage{
is_even(nums)
}
\arguments{
\item{nums}{numeric}
}
\value{
logical
}
\description{
Check if number(s) is/are even
}
\examples{
is_even(3)
is_even(c(4,5))
}
