% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{Check if a number or numbers is/are odd.}
\usage{
is_odd(n)
}
\arguments{
\item{n}{numeric}
}
\value{
logical
}
\description{
Check if a number or numbers is/are odd.
}
\examples{
is_odd(1)
is_odd(2)
is_odd(c(3,4))
}
