% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels_length.R
\name{levels_length}
\alias{levels_length}
\title{Check the length of the factor against the number of levels}
\usage{
levels_length(x)
}
\arguments{
\item{x}{factor}
}
\value{
logical
}
\description{
This function detects factors that should be character because # unique values = length
This function does not convert these problematic factors to character.
}
\examples{
freq_out(iris$Species)
levels_length(iris$Species)
}

