% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighthouses.R
\docType{data}
\name{lighthouses}
\alias{lighthouses}
\title{Historical light aids to navigation 1514-1911}
\format{
A data frame with 603 rows and 15 variables:
\describe{
\item{FID}{unique ID}
\item{Name}{unique name of each lighthouse}
\item{LH_MED}{number of individual lights possessed in medieval period}
\item{LH_1693}{number of individual lights possessed in 1693}
\item{LH_1753}{number of individual lights possessed in 1753}
\item{LH_1831}{number of individual lights possessed in 1831}
\item{REACH_1831}{night-time visibility range in 1831 (miles)}
\item{LH_1851}{number of individual lights possessed in 1851}
\item{REACH_1851}{night-time visibility range in 1851 (miles)}
\item{LH_1911}{number of individual lights possessed in 1911}
\item{REACH_1911}{night-time visibility range in 1911 (miles)}
\item{Height}{height, in feet}
\item{YearBuilt}{first recorded year of construction}
\item{POINT_X}{geographical coordinate X of the lighthouse}
\item{POINT_Y}{geographical coordinate Y of the lighthouse}
...
}
}
\source{
\url{https://reshare.ukdataservice.ac.uk/854172/}
}
\usage{
lighthouses
}
\description{
A dataset containing information on historical lighthouses in England and Wales.
}
\keyword{datasets}
