% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowercase.R
\name{lowercase}
\alias{lowercase}
\title{change de the factor to lowercase letters}
\usage{
lowercase(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
change de the factor to lowercase letters
}
\examples{
lowercase(iris$Species)
}
