% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_dif.R
\name{mean_dif}
\alias{mean_dif}
\title{Difference of means of two numeric variables}
\usage{
mean_dif(x, y)
}
\arguments{
\item{x}{numeric, of any length}

\item{y}{numeric, of any length}
}
\value{
numeric
}
\description{
Takes the mean of two numeric variables (x and y) then takes the difference between them (x-y).
}
\examples{
#' mean_dif(mtcars$cyl,mtcars$gear)
}

