% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitracks.R
\name{multitracks}
\alias{multitracks}
\title{Constructor for an object of a class \code{multitracks}}
\usage{
multitracks(lineobject)
}
\arguments{
\item{lineobject}{Object of a class \code{mytrack}}
}
\value{
Object of a class \code{multitracks} which holds multiple \code{mytrack} objects
}
\description{
Constructor for an object of a class \code{multitracks}
}
\examples{
\dontrun{
my_tracks <- multitracks(line1)
}
}
