% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_factors.R
\name{new_factor}
\alias{new_factor}
\title{Set levels of a factor to the order appeared in data}
\usage{
new_factor(x)
}
\arguments{
\item{x}{a factor or a vector}
}
\value{
a factor with levels in the order/reverse order appeared in x.
}
\description{
Set levels of a factor to the order appeared in data
}
\examples{
new_factor(c("d","b","a","d"))
new_factor_rev(c("a","b","c","a"))
}
