% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_factors.R
\name{new_factor_rev}
\alias{new_factor_rev}
\title{Set levels of a factor to the reverse order appeared in data}
\usage{
new_factor_rev(x)
}
\arguments{
\item{x}{a factor}
}
\value{
a factor with levels in the reverse order appeared in x.
}
\description{
Set levels of a factor to the reverse order appeared in data
}
\examples{
new_factor_rev(factor(c("a","b","c","a")))
}
