% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_reorder.R
\name{new_reorder}
\alias{new_reorder}
\title{A new version of reorder that uses desc}
\usage{
new_reorder(x)
}
\arguments{
\item{x}{a factor or a vector}
}
\value{
a factor
}
\description{
This function give you the ability to sort your factors in descending order.
}
\examples{
new_reorder(factor(c("c","b","a")))
}
