% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfactors.R
\name{newrev}
\alias{newrev}
\title{Set the levels reversed}
\usage{
newrev(x)
}
\arguments{
\item{f}{a factor}
}
\value{
a factor with levels in the reverse order appeared in x.
}
\description{
Set levels of a factor to the reversed order in which they appear in the data.
}
\examples{
newrev(factor(c("s","t","a","t")))
}
