% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\docType{data}
\name{numbers}
\alias{numbers}
\title{numbers data}
\format{
A data frame with 10 rows and 1 variables:
\describe{
\item{number}{Number}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
numbers
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
