% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_asis.R
\name{order_asis}
\alias{order_asis}
\title{Order Factors Based on order in Data}
\usage{
order_asis(a)
}
\arguments{
\item{a}{vector}
}
\value{
factor
}
\description{
Ever just wish R would order things in the order they appear? I do. I also wish factors made some friggen sense.
I can't help factors make anymore sense, but this function will order factors in the order they appear in the data!
}
\examples{
#' order_asis(mtcars$cyl)
}

