% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdesc.R, R/fordered.R
\name{fdesc}
\alias{fdesc}
\alias{fordered}
\title{Reorder a factor in descending order}
\usage{
fdesc(x)

fordered(x)
}
\arguments{
\item{x}{factor}
}
\value{
fdesc(x) returns a factor in descending order

fordered(x) returns ordered levels
}
\description{
This function is used to reorder factor in descenging order, it will come to an error when the param is not a factor.

fordered(x) is the function used to sets levels to the order in which they appear in the data, the input should be a factor, otherwise, the function will come to an error.
fdesc(x) is used to reorder factor in descenging order, it will come to an error when the param is not a factor.
}
\examples{
fdesc(factor(c("a","b","c"))) # Levels: c b a
fdesc(iris$Species) # Levels: virginica versicolor setosa
fordered(iris$Species) # Levels: setosa < versicolor < virginica
}
