% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownercat_data.R
\docType{data}
\name{ownercat_data}
\alias{ownercat_data}
\title{Owner Cat Data}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{name}{Owner's Name}
\item{number}{Cat's Name}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
ownercat_data
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
