% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oztokg_data.R
\docType{data}
\name{oztokg_data}
\alias{oztokg_data}
\title{oztokg example data}
\format{
A data frame with 20 rows and 1 variable1:
\describe{
\item{value}{example number in unit oz to convert to kg}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
oztokg_data
}
\description{
oztokg example data
}
\keyword{datasets}
