% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{plot.mytrack}
\alias{plot.mytrack}
\title{Plot method for class "mytrack"}
\usage{
\method{plot}{mytrack}(x, y, ...)
}
\arguments{
\item{x}{Object of a class "mytrack"}

\item{y}{Not used}

\item{...}{Not used}
}
\value{
Plot in Cartesian xy system where Longitude is plotted on x and
latitude is plotted on y
}
\description{
Plot method for class "mytrack"
}
\examples{
\dontrun{
plot(obj1)}
}
