% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitracks.R
\name{print.multitracks}
\alias{print.multitracks}
\title{\code{\link{print}} method for class \code{multitracks}}
\usage{
\method{print}{multitracks}(x, ...)
}
\arguments{
\item{x}{object of class \code{multitracks}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
prints a list of all tracks in the container
}
\description{
\code{\link{print}} method for class \code{multitracks}
}
\examples{
\dontrun{
print(my_tracks)
}
}
