% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{print.mytrack}
\alias{print.mytrack}
\title{Print method for class "mytrack"}
\usage{
\method{print}{mytrack}(x, ...)
}
\arguments{
\item{x}{object of a class "mytrack"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints elements of object of a class "mytrack"
}
\description{
Print method for class "mytrack"
}
\examples{
\dontrun{
print(obj1)}
}
