% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PS06.R
\docType{data}
\name{PS06}
\alias{PS06}
\title{Medicaid Expenditure 2019}
\format{
a data frame with 51 rows and 4 variables:
\describe{
\item {STATEFP}{STATEFP}
\item {TtlMedExpC}{Total Medicaid spending in character format}
\item {TtlMedExpN}{Total Medicaid spending in numeric format}
\item {Year}{Year of the data}
}
}
\source{
\url{https://www.kff.org/medicaid/state-indicator/total-medicaid-spending/?currentTimeframe=0&sortModel=\%7B\%22colId\%22:\%22Location\%22,\%22sort\%22:\%22asc\%22\%7D}
}
\usage{
PS06
}
\description{
A data set containing medicaid expenditure data at the state level for 2019.
}
\keyword{datasets}
