% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sample_table.R
\name{random_sample_table}
\alias{random_sample_table}
\title{Sample Table}
\usage{
random_sample_table(x, nsample = 5)
}
\arguments{
\item{x}{data.frame}

\item{nsample}{number of rows to sample from tbl}
}
\value{
data.frame
}
\description{
Pull out a random sample of rows as a data.frame
}
\examples{
# randomly sample rows from a table (by default 10)
peek(iris)
# randomly sample 10 rows from a table
peek(iris, 10)
}

