% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df.R
\name{read_df}
\alias{read_df}
\title{Read a data frame}
\usage{
read_df(filepath)
}
\arguments{
\item{filepath}{string file name and path}

\item{df}{dataframe}
}
\description{
Read a data frame in a tab delimited file while retaining factor levels.
}
