% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder Levels of a factor}
\usage{
reorder(x, X, FUN = mean, ..., order = is.ordered(x))
}
\arguments{
\item{x}{an atomic factor vector}

\item{X}{a vector of the same length as x, whose subset of values for each unique level of x 
determines the eventual order of that level.}

\item{FUN}{a function whose first argument is a vector and returns a scalar, to be applied to 
each subset of X determined by levels of x}

\item{...}{optional extra arguments to be supplied to FUN}

\item{order}{a logical defining whether the output is ordered or not}
}
\value{
A factor or an ordered vector, with the order of levels determined by FUN applied to X
grouped by x. The levels are sorted in descending order of the values returned by FUN.
}
\description{
A new version of reorder that uses desc from dplyr
}

