% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder2.R
\name{reorder2}
\alias{reorder2}
\title{Reorder the levels of the first argument, which is a vairable in a dataframe based on 
the values of the second argument in decreasing order.}
\usage{
reorder2(df, a, b)
}
\arguments{
\item{df}{dataframe}

\item{a}{variable}

\item{b}{variable}
}
\value{
dataframe
}
\description{
Reorder the levels of the first argument, which is a vairable in a dataframe based on 
the values of the second argument in decreasing order.
}
\examples{
reorder2(mtcars, mtcars$cyl, mtcars$hp) 
reorder2(mtcars, mtcars$cyl, mtcars$wt)
reorder2(iris, iris$Species, iris$Sepal.Length)
}
