% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder a factor using desc from dplyr}
\usage{
reorder_desc(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor; gives reordered levels in decending alphabetical order
}
\description{
Reorder a factor using desc from dplyr
}
\examples{

x <- factor(c("a", "b", "c", "d"))
levels(x)
levels(reorder_desc(x))
}
