% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_descending.R
\name{reorder_descending}
\alias{reorder_descending}
\title{Reorder levels in descending order}
\usage{
reorder_descending(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
A version of reorder that uses dplyr::desc() to reorder
factor levels in descending order
}
\examples{
data_2 <- factor(c("setosa","versicolor"))
reorder_descending(data_2)
}
