% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_factor.R
\name{reorder_factor}
\alias{reorder_factor}
\title{Reorder factor levels in descending order}
\usage{
reorder_factor(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Reorder factor levels in descending order
}
\examples{
reorder_factor(iris$Species)
}

