% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_new.R
\name{reorder_new}
\alias{reorder_new}
\title{Reorder levels in descending order}
\usage{
reorder_new(x)
}
\arguments{
\item{x}{factor}
}
\value{
This reorder fuction is using desc() in dplyr to reorder levels in descending order
}
\description{
Reorder levels in descending order
}
\examples{
reorder_new(factor(c("1","2","3")))
}
