% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room_data.R
\docType{data}
\name{room_data}
\alias{room_data}
\title{Cushing House dorm rooms and features}
\format{
A data frame with 51 rows and 8 variables:
\describe{
\item{Hall}{house name}
\item{Floor}{floor number}
\item{Room}{room number}
\item{Room Type}{room type}
\item{SouthwestFacing}{presence of southwest-facing window(s)}
\item{Corner}{room located in corner}
\item{NotNearBathroom}{room not located near bathroom}
...
}
}
\source{
\url{https://www.smith.edu/student-life/residence-life}
}
\usage{
room_data
}
\description{
A dataset containing a hand-picked selection of prime dorm rooms in Cushing House
at Smith College and noteable features like whether or not they are southwest-facing,
in a corner, or near a bathroom
}
\keyword{datasets}
