% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roomrank.R
\name{roomrank}
\alias{roomrank}
\title{Rank dorm rooms}
\usage{
roomrank(room_df)
}
\arguments{
\item{room_df}{room dataframe}
}
\value{
A tibble
}
\description{
Ranks a data frame of a selection of Smith College dorm rooms from most
favorable to least favorable based on which of the following features it
has: southwest-facing, in a corner, and not near a bathroom.
}
\examples{
roomrank(room_data)
}
