% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayhello.R
\name{sayhello}
\alias{sayhello}
\title{a function for say hello}
\usage{
sayhello(txt = "hello")
}
\arguments{
\item{txt}{char}
}
\description{
a function for say hello
}
\examples{
sayhello('hwb')
sayhello('hello,hwb')
}
\author{
webber
}
