% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_size.R
\docType{data}
\name{schoolsize_data}
\alias{schoolsize_data}
\title{School size data}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{name}{School Name}
\item{size}{School size}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
schoolsize_data
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
