% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_factor.R
\name{sort_factor}
\alias{sort_factor}
\title{Sort the factor in descending order}
\usage{
sort_factor(a)
}
\arguments{
\item{a}{input}
}
\value{
the descending sorted order factor
}
\description{
This is a function is used to sort the input facotr into descending order
}
\examples{
sort_factor(factor(c("hits", "hits", "your", "eyeballs")))
}
