% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_code_filter.R
\name{stem_code_filter}
\alias{stem_code_filter}
\title{Select data using stem code}
\usage{
stem_code_filter(df, stem_code, include = TRUE)
}
\arguments{
\item{df}{data frame}

\item{stem_code}{character or list}

\item{include}{logical; TRUE will filter the data frame to include only entries containing the given stem code(s) and FALSE will remove all entries containing the given code(s)}
}
\value{
data frame
}
\description{
Filters ForestGeo's SCBI tree census data by stem code. Allows user to select for or exclude entries based on the presence of user-defined stem code(s). Please reference the following website for stem code definitions: \url{https://github.com/SCBI-ForestGEO/SCBI-ForestGEO-Data/blob/master/tree_main_census/metadata/SCBI_stem_codes.pdf}.
}
\examples{
stem_code_filter(treecensus, "F", include = FALSE)
my_codes <- c("A", "M", "DC", "D", "B")
stem_code_filter(treecensus, my_codes)

}
