% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetrack.R
\name{storetrack}
\alias{storetrack}
\title{Creates new object of a class "track".}
\usage{
storetrack(my_name, timestamp, coordinate_system, coords)
}
\arguments{
\item{my_name}{Name of track of a "character" type}

\item{timestamp}{Vector holding timestamps of a "character" type}

\item{coordinate_system}{EPSG number for datum and projection}

\item{coords}{Data frame with longitude and latitude}
}
\value{
Object of a type "track"
}
\description{
Creates new object of a class "track".
}
\examples{
## Create vector with timestamps
my_timestamps <- c("2009-12-25 18:39:21 CST","2009-12-25 18:39:31 CST","2009-12-25 18:39:41 CST")

## Create data frame with coordinates
my_coordinates <- data.frame("longitude"= c(22.23,34.21,45.34),"latitude" = c(33.16,12.32,22.21))

## Create object of class "track"
storetrack("Track1",my_timestamps,4326,my_coordinates)
}
