% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{storetracks3}
\alias{storetracks3}
\title{Constructor for object of a type "mytrack"}
\usage{
storetracks3(name, time, crs, coordinates)
}
\arguments{
\item{name}{Name of the track}

\item{time}{Timestamps for points}

\item{crs}{CRS code}

\item{coordinates}{LON-LAT coordinates}
}
\value{
Object of a class "mytrack
}
\description{
Constructor for object of a type "mytrack"
}
\examples{
\dontrun{
obj1 <- storetracks3("Track1", my_timestamps, 4326, my_coordinates)
}
}
