% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Newthon-Raphson}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{its an ex}

\item{split}{its split}
}
\value{

}
\description{
Newthon-Raphson
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit(x, split = ",")
}
