% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfactor.R
\name{subfactor}
\alias{subfactor}
\title{First three characters of factor}
\usage{
subfactor(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
This function returns a factor with just the first three letters of each element.
This may help to not have all the complete names when plotting. Note: There is a problem 
sometimes with levels that are not represented by only 3 letters like United States
in the case of countries
}
\examples{
subfactor(iris$Species)
}
