% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitracks.R
\name{summary.multitracks}
\alias{summary.multitracks}
\title{\code{\link{summary}} method for class \code{multitracks}}
\usage{
\method{summary}{multitracks}(object, ...)
}
\arguments{
\item{object}{object of class \code{multitracks}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
gives summary of all tracks in the container
}
\description{
\code{\link{summary}} method for class \code{multitracks}
}
\examples{
\dontrun{
summary(my_tracks)
}
}
