% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{name}{Name}
\item{number}{Favorite number}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
toy_data
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
