% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetrack.R
\docType{class}
\name{track-class}
\alias{track-class}
\title{An S4 class for holding coordinates of a track}
\description{
An S4 class for holding coordinates of a track
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of track}

\item{\code{time}}{Timestamps for points}

\item{\code{crs}}{Code for CRS}

\item{\code{coordinates}}{Data frame holding coordinates}
}}

