% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecensus.R
\docType{data}
\name{treecensus}
\alias{treecensus}
\title{SCBI Tree Census}
\format{
a dataframe with 72555 rows and 17 variables; please visit source for variable definitions
}
\source{
\url{https://github.com/SCBI-ForestGEO/SCBI-ForestGEO-Data}
}
\usage{
treecensus
}
\description{
Data from the 2018 census of trees in the SCBI ForestGEO Plot.
}
\keyword{datasets}
