% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uppercase.R
\name{uppercase}
\alias{uppercase}
\title{change de the factor to capital letters}
\usage{
uppercase(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
change de the factor to capital letters
}
\examples{
uppercase(iris$Species)
}
