% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbottle_data.R
\docType{data}
\name{waterbottle_data}
\alias{waterbottle_data}
\title{Water Bottle data}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{name}{Name}
\item{bottle_type}{Water Bottle Type}
...
}
}
\source{
\url{https://r-pkgs.org/data.html#data-data/}
}
\usage{
waterbottle_data
}
\description{
Based on R Packages Chapter 14
}
\keyword{datasets}
