% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df.R
\name{write_df}
\alias{write_df}
\title{Write a data frame}
\usage{
write_df(df, filepath)
}
\arguments{
\item{df}{dataframe}

\item{filepath}{string file name and path}
}
\description{
Write a daat frame in a tab delimited file while retaining factor levels.
}
