% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf.R
\name{writedf}
\alias{writedf}
\title{Read data frames from plain text delimited files}
\usage{
writedf(df, filepath)
}
\arguments{
\item{df}{data frame}

\item{filepath}{charactor}
}
\value{
charactor filename
}
\description{
Read data frames from plain text delimited files
}
