% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_col.R
\name{add_col}
\alias{add_col}
\title{Adds column}
\usage{
add_col(x, name, value, where = ncol(x) + 1)
}
\arguments{
\item{x}{dataframe}

\item{name}{name of the column}

\item{value}{value}

\item{where}{position}
}
\value{
dataframe
}
\description{
Adding a column to the dataframe
}
\examples{

df <- data.frame(x = 1)
add_col(df, "y", 2)

}
