% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{backward}
\alias{backward}
\title{Backwards factors}
\usage{
backward(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Create a new factor with the inverse order of the original factor
}
\examples{
backward(iris$Species)
}
