% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fac.R
\name{detect_fac}
\alias{detect_fac}
\title{Detect factors that should be characters}
\usage{
detect_fac(a)
}
\arguments{
\item{a}{list of factors}
}
\value{
a list of characters
}
\description{
Detect factors that should be characters
}
\examples{
detect_fac(factor(c("a","b","c")))
}

