% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_factors.R
\name{detect_factors}
\alias{detect_factors}
\title{Title}
\usage{
detect_factors(x)
}
\arguments{
\item{x}{a vector or a factor}
}
\description{
Title
}
\examples{
detect_factors(c("a","b","a"))
detect_factor(c("a","b","c"))
}
