% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fct.R
\name{detect_fct}
\alias{detect_fct}
\title{Detect whether a factor is truly factor or character}
\usage{
detect_fct(x)
}
\arguments{
\item{x}{factor}
}
\value{
boolean; \code{TRUE} if number of unique values does not equal to length of factor,
   \code{FALSE} otherwise and the input should be a character
}
\description{
Detect whether a factor is truly factor or character
}
\examples{
a <- factor(c("high", "high", "low"))
b <- factor(c("high", "low", "medium"))

This should be a factor:
detect_fct(a)

This should be a character:
detect_fct(b)

}
