% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fct.R
\name{detect_fct}
\alias{detect_fct}
\title{Detect Factors}
\usage{
detect_fct(x)
}
\arguments{
\item{x}{factor}
}
\value{
(logical) \code{TRUE} or \code{FALSE}. \code{TRUE} if the length of x is not equal to the length of \code{unique(x)};
otherwise, \code{FALSE}
}
\description{
Detect factors that should be character because the number of unique values of a factor is equal to its length
}
\examples{
a <- as.factor(c("yes", "yes", "no", "yes"))
detect_fct(a)  # TRUE
b <- as.factor(c("Apple", "Banana", "Orange", "Cherry"))
detect_fct(b)  # FALSE

}
