% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_asis.R
\name{factor_asis}
\alias{factor_asis}
\title{Encodes a vector as a factor based on the same levels (as is)}
\usage{
factor_asis(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Encodes a vector as a factor based on the same levels (as is)
}

