% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_asis.R
\name{factor_asis}
\alias{factor_asis}
\title{Factor Asis}
\usage{
factor_asis(x)
}
\arguments{
\item{x}{a vector of data}
}
\value{
a factor
}
\description{
Encode a vector as a factor with the order in which they appear in the data
}
\examples{
x <- c(1, 2, 3, 14, 0, 7)
factor_asis(x)
y <- c("a", "c", "d", "e", "b")
factor_asis(y)

}
