% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbinbd.R
\name{fbind}
\alias{fbind}
\title{Bind two factors
#' Bind two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor a}

\item{b}{factor b}
}
\value{
a new factor that binds a and b
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
