% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(a, b, ...)
}
\arguments{
\item{a}{factor}

\item{b}{factor}

\item{...}{other arguments}
}
\value{
factor
}
\description{
Create a new factor from two existing factors, where the new
factor's levels are the union of the levels of the input
factors.
}
\examples{
fbind(factor(letters[1:3]), factor(letters[26:24]))
}
