% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor 1}

\item{b}{factor 2}
}
\value{
combined factor 1 & 2
}
\description{
Bind two factors
}
\examples{
fbind(as.factor(c("a","b","c")), as.factor(c("d","e","f")))
}
