% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors in a sensible way}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{a factor}

\item{b}{a factor}
}
\value{
a factor
}
\description{
Bind two factors in a sensible way
}
\examples{
a <- factor("character", "hits", "your", "eyeballs")
b <- factor("but", "integer", "where it", "counts")
fbind(a, b)
}
