% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Combine two factors including levels}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{`factor`}

\item{b}{`factor`}
}
\value{
c `factor`, combined a and b including levels
}
\description{
Combine two factors including levels
}
\examples{
\dontrun{fbind(iris$Species, Pyromycin$state)}

}
