% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind factors}
\usage{
fbind(x, y)
}
\arguments{
\item{x}{First factor.}

\item{y}{Second factor.}
}
\value{
A new factor, with the levels of `x` and `y`.
}
\description{
Combines factors, like you would expect.
}
\examples{
a <- factor(c("hola", "hello", "salut"))
b <- factor(c("adios", "bye", "salut"))
fbind(a, b)

}
