% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Add roxygen comments by
# bind two factors
# this is called the 'description':
Add roxygen comments by putting cursor inside function and go to
code -> insert roxygen skeleton}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A factor.}

\item{b}{A factor.}
}
\value{
A factor.
}
\description{
Add roxygen comments by
# bind two factors
# this is called the 'description':
Add roxygen comments by putting cursor inside function and go to
code -> insert roxygen skeleton
}
\examples{
#(a <- factor(c("character", "hits", "your", "eyeballs")))
}
