% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{factor}{}
}
\value{
factor
}
\description{
Bind two factors together
}
\examples{
fbind(factor(letters[1:3]),factor(letters[26:24]))
}
