% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Combine two factors, including levels}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{\code{factor}}

\item{b}{\code{factor}}
}
\value{
\code{factor}, combined \code{a} and \code{b} including levels
}
\description{
Combine two factors, including levels
}
\examples{
fbind(iris$Species, Puromycin$state)

}
