% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Binds two lists of factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{A list of factors}

\item{b}{A list of factors}
}
\value{
A list of factors containing the factors of a and b
}
\description{
Binds two lists of factors
}
\examples{
a <- factor(c("character", "hits", "your", "eyeballs"))
b <- factor(c("but", "integer", "where it", "counts"))
fbind(a, b)
}
