% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Concatenate factors, including levels}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{\code{factor}}

\item{b}{\code{factor}}
}
\value{
\code{factor}, concatenation of \code{a} and \code{b}
}
\description{
Concatenates two factors with two differerent sets of levels, returning
a factor with levels that are the union of the input factors' levels.
}
\examples{
fbind(iris$Species, Puromycin$state)

}
