% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Function to join factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{first factor}

\item{b}{second factor}
}
\value{
factor
}
\description{
some description to add into the package
}
\examples{
a <- c("first", "second")
b <- c("one", "three")
fbind(a, b)

}
