% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{your title for the function}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor 1}

\item{b}{factor 2}
}
\value{
facotr combined
}
\description{
description here
}
\examples{
a = factor(c(1,2,3))
b = factor(c(2,3,4))
fbind(a,b)
}
\references{
blah blah
}
