% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcheck.R
\name{fcheck}
\alias{fcheck}
\title{Check the input}
\usage{
fcheck(x)
}
\arguments{
\item{x}{factor}
}
\value{
Nothing, if it comes to an error, it will stop
}
\description{
Check whether the input is a factor or not, if it's not a factor, then the function wil throw an error.
}
\examples{
fcheck(iris$Species)
}
