% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcheck.R
\name{fcheck}
\alias{fcheck}
\title{Check for input is a a valid factor}
\usage{
fcheck(x)
}
\arguments{
\item{x}{a factor.}
}
\value{
An error is raised if x is not a factor
}
\description{
Check for input is a valid factor or not
}
