% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fclean.R
\name{fclean}
\alias{fclean}
\title{Clean factor levels
For all levels, remove white spaces and convert to title case.}
\usage{
fclean(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Clean factor levels
For all levels, remove white spaces and convert to title case.
}
\examples{
production <- factor(c("  HIGH production", "MED production  ",  "MED production  ", " low Production"))
fclean(production)
}
