% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcount.R
\name{fcount}
\alias{fcount}
\title{make a sorted frequency table for a factor}
\usage{
fcount(x)
}
\arguments{
\item{x}{a factor}
}
\value{
A tibble
}
\description{
will count the number of occurances for each
factor level
}
\examples{
fcount(iris$Species)
}
