% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdetect.R
\name{fdetect}
\alias{fdetect}
\title{Detect factors that should be character}
\usage{
fdetect(x)
}
\arguments{
\item{x}{a factor. If x is not a factor, an error is raised.}
}
\value{
Returns boolean
Returns \code{TRUE} if # unique values = length
An error is raised if # unique values != length
}
\description{
Detect factors that should be character because # unique values = length
}
\examples{
fdetect(factor(c("a", "b","a"))) # FALSE
fdetect(factor(c("a", "b", "c"))) # TRUE
}
