% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgap_rw.R
\name{fgap_rw}
\alias{fgap_rw}
\title{This function shows how takes the `gapmider` excerpt data frame and 
compares `write.table()` and `saveRDS()` functions on mini data frame that is the result of the `fgap_coefs()`.}
\usage{
fgap_rw(dat, na.rm = TRUE)
}
\arguments{
\item{dat}{gapminder data frame}

\item{na.rm}{removes NA (if any)}
}
\value{
txt and rds files and a mini data frame
}
\description{
This function shows how takes the `gapmider` excerpt data frame and 
compares `write.table()` and `saveRDS()` functions on mini data frame that is the result of the `fgap_coefs()`.
}
\examples{
fgap_rw(gapminder::gapminder)
}

