% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.R
\docType{package}
\name{foofactors-package}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: An example for the useR! conference}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Vincent ROCHER \email{vincent.rocher@univ-tlse3.fr}

}
\keyword{internal}
