% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/foofactors-package.r
\docType{package}
\name{foofactors}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: A project for stat545 2015}
\description{
Here I do some homework to practise producing a package
I produced a function that combines factors' levels a the whim of the user.
Then I spent a heck of a lot of time trying to turn it into a package, based off a package initiated by Jenny.
}

