% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.r
\docType{package}
\name{foofactors}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: Make factors less annoying}
\description{
Factors have driven people to extreme measures, like ordering custom
conference ribbons and laptop stickers to express how we feel about
factors. And yet, sometimes you need them.
This package is designed to make them less annoying.
}

