% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.r
\docType{package}
\name{foofactors}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: Manipulate factors as you will}
\description{
This package performs functions on factors, like binds 2 factors into one.
}

