% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foofactors-package.r
\docType{package}
\name{foofactors}
\alias{foofactors}
\alias{foofactors-package}
\title{foofactors: Make factors less annoying}
\description{
Here I will continue adding functions into the foofactor package Jenny gave us.
}

