% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forder.R
\name{forder}
\alias{forder}
\title{Reorder Factor Levels}
\usage{
forder(df, factor, var)
}
\arguments{
\item{var}{}
}
\value{
factor
}
\description{
Reorder Factor Levels
}
\examples{
iris$Species <- forder(iris, iris$Species, iris$Sepal.Width)
boxplot(Sepal.Width ~ Species, data = iris)
}

