% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freorder.R
\name{freorder}
\alias{freorder}
\title{Reorder Levels of a Factor}
\usage{
freorder(a, b)
}
\arguments{
\item{a}{An atomic vector, usually a factor (possibly ordered). The vector is treated as a categorical variable
whose levels will be descendingly reordered.}

\item{b}{a vector of the same length as a, whose subset of values for each unique level of a determines the eventual
order of that level. b is required to contain numeric values.}
}
\value{
A factor with the order of the levels determined by applying the numeric order of b grouped by a. The levels
are ordered such that the values returned are in descending order.
}
\description{
freorder is a generic function. The "default" method treats its first argument as a categorical variable,
and reorders its levels in descending order based on the values of a second variable, usually numeric.
}
\examples{
freorder(school <- c("School A", "SChool B", "School C", "School D", "School E", "School F"), N <- c(25,28,12,22,30,25))
}

