% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_out.R
\name{freq_out}
\alias{freq_out}
\title{Make a frequency table for a factor}
\usage{
freq_out(x)
}
\arguments{
\item{x}{factor}
}
\value{
tbl_df
}
\description{
Use this function to count the number of entries at each level of the factor.
}
\examples{
#' freq_out(iris$Species)
}

