% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fset.R
\name{fset}
\alias{fset}
\title{Sets levels of a factor}
\usage{
fset(x)
}
\arguments{
\item{x}{a factor. If x is not a factor, an error is raised.}
}
\value{
This function returns a factor with the level set to the order in which they appear in the data.
}
\description{
Sets levels to the order in which they appear in the data
}
\examples{
fset(factor(c("B", "A", "D"))) # Levels: B A D

}
