% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_char.R
\name{is_char}
\alias{is_char}
\title{Should a factor be character?}
\usage{
is_char(x)
}
\arguments{
\item{x}{}
}
\value{
boolean
}
\description{
Detect factors that should be character because # unique values = length
}
\examples{
is_char(factor(c("a", "b", "c", "d")))
}

