% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowercase.R
\name{lowercase}
\alias{lowercase}
\title{Lowercase factor}
\usage{
lowercase(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Returns the same factor but with all lowercase letters
}
\examples{
lowercase(iris$Species)
}
