% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_factors.R
\name{new_factor}
\alias{new_factor}
\title{Set levels of a factor to the order appeared in data}
\usage{
new_factor(x)
}
\arguments{
\item{x}{a factor}
}
\value{
a factor with levels in the order appeared in x.
}
\description{
Set levels of a factor to the order appeared in data
}
\examples{
new_factor(factor(c("d","b","a","d")))
}
