% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_reorder.R
\name{new_reorder}
\alias{new_reorder}
\title{A new version of reorder() that uses desc()}
\usage{
new_reorder(x)
}
\arguments{
\item{x}{a factor or a vector}
}
\value{
a factor with levels in the order/reverse order appeared in input.
}
\description{
A new version of reorder() that uses desc()
}
\examples{
new_reorder(c("c","b","a"))
}
