% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks.R
\name{plot.mytrack}
\alias{plot.mytrack}
\title{\code{\link{plot}} method for class \code{mytrack}}
\usage{
\method{plot}{mytrack}(x, y, ...)
}
\arguments{
\item{x}{Object of a class \code{mytrack}}

\item{y}{Not used}

\item{...}{Not used}
}
\value{
Plot in Cartesian xy system where Longitude is plotted on x and
latitude is plotted on y
}
\description{
\code{\link{plot}} method for class \code{mytrack}
}
\examples{
\dontrun{
plot(obj1)}
}
