% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks.R
\name{print.mytrack}
\alias{print.mytrack}
\title{\code{print} method for class \code{mytrack}}
\usage{
\method{print}{mytrack}(x, ...)
}
\arguments{
\item{x}{object of a class \code{mytrack}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints elements of object of a class \code{mytrack}
}
\description{
\code{print} method for class \code{mytrack}
}
\examples{
\dontrun{
print(obj1)}
}
