% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\docType{data}
\name{PS06}
\alias{PS06}
\title{Medicaid Expenditure 2019.}
\format{
A dataframe with 51 rows and 4 variables:
}
\source{
\url{https://github.com/GeoDaCenter/opioid-policy-scan/blob/master/data_final/PS06_2019_S.csv}
}
\usage{
PS06
}
\description{
A dataset containing medicaid expenditure data in 2019
}
\keyword{datasets}
