% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder Factors}
\usage{
reorder(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Reorder factors with decreasing order.
}
\examples{
a <- as.factor(c(21,23,45,64,12,22,3, 4, 4, 4, 4))
reorder(a)
b <- as.factor(c("apple", "orange", "banana", "cherry", "orange"))
reorder(b)

}
