% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder Factors w/descending option.}
\usage{
reorder_desc(a)
}
\arguments{
\item{a}{factor}
}
\value{
factor
}
\description{
Hate having to list all of the factors in reverse order when you JUST WANT TO HAVE THEM IN DESCENDING ORDER!?!
Look no further friend! This function will cure was ails you, feed the poor, solve the millenium problems....Well, not really.
But it does give you the ability to sort your factors in descending order.
}
\examples{
#' chr_factor(factor(c('BumbleBees', 'Bees', 'Bears', 'Creates', 'Capital', 'For','Unicorns'))

}

