% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder vector in descending order}
\usage{
reorder_desc(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Create a new factor with with a descending order of the levels of the original factor
}
\examples{
reorder_desc(iris$Species)
}
