% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder factors using desc function.}
\usage{
reorder_desc(x, X, FUN = mean, ...)
}
\arguments{
\item{x}{factor}

\item{X}{factor}

\item{FUN}{function}
}
\value{
factor
}
\description{
Reorder factors using desc function.
}
