% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfactor.R
\name{subfactor}
\alias{subfactor}
\title{Extract first three characters of factor}
\usage{
subfactor(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Extract first three characters of factor
}
\examples{
subfactor(iris$Species)
}
