% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks.R
\name{summary.mytrack}
\alias{summary.mytrack}
\title{\code{\link{summary}} method for class \code{mytrack}}
\usage{
\method{summary}{mytrack}(object, ...)
}
\arguments{
\item{object}{object of a class \code{mytrack}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary of objects of a class \code{mytrack}
}
\description{
\code{\link{summary}} method for class \code{mytrack}
}
\examples{
\dontrun{
summary(obj1)}
}
