% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storetracks3.R
\name{summary.mytrack}
\alias{summary.mytrack}
\title{Summary method for class "mytrack"}
\usage{
\method{summary}{mytrack}(object, ...)
}
\arguments{
\item{object}{object of a class "mytrack"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints summary of objects of a class "mytrack"
}
\description{
Summary method for class "mytrack"
}
\examples{
\dontrun{
summary(obj1)}
}
